+++ docs/conf.py
@@ -27,7 +27,7 @@ author = 'C. W.'
 # The short X.Y version
 version = '0.6.7'
 # The full version, including alpha/beta/rc tags
-release = '0.6.6'
+release = '0.6.7'
 
 # -- General configuration ---------------------------------------------------