+++ moban/plugins/copy.py
@@ -21,6 +21,9 @@ class ContentForwardEngine(object):
     templating mechanism.
     """
 
+    ACTION_IN_PRESENT_CONTINUOUS_TENSE = "Copying"
+    ACTION_IN_PAST_TENSE = "Copied"
+
     def __init__(self, template_fs, extensions=None):
         self.template_fs = template_fs